now very ready to refactor
